[scmd] refactor split scmd source file into three.
